# No dictionary group by perf